feat: refactor clubs
