+++ src/pages/club/utils/covertLinks.ts
@@ -1,0 +1,7 @@
+export const convertLinks = (text: string) => {
+  const urlRegex = /(https?:\/\/[^\s]+)/g;
+  return text.replace(
+    urlRegex,
+    '<a href="$1" target="_blank" rel="noopener noreferrer">$1</a>'
+  );
+};
